requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dataclasses==0.6
Django==5.0.6
django-oauth-toolkit==3.0.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.2.31
dnspython==2.7.0
idna==3.10
jwcrypto==1.5.6
mongoengine==0.29.1
oauthlib==3.2.2
pillow==11.0.0
pycparser==2.22
pymongo==4.10.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
